FT2-1503 Round up Min-Max hours and remove special characters from URLs for JP

+++ DFC.Digital/DFC.Digital.Data/Model/OrchardCore/OcJobProfile.cs
@@ -167,12 +167,12 @@ namespace DFC.Digital.Data.Model.OrchardCore
 
     public class Minimumhours
     {
-        public float? Value { get; set; }
+        public int? Value { get; set; }
     }
 
     public class Maximumhours
     {
-        public float? Value { get; set; }
+        public int? Value { get; set; }
     }
 
     public class HiddenalternativetitleIds
